Fix unicode-preference
- unicode_prefernces bug fixed
- should now set user pref correctly
