+++ tests/conftest.py
@@ -10,7 +10,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..',
 
 # Imports
 import bot.setup.register_commands as register_commands
-from bot.commands import bot_settings, debt_display, debt_management
+from bot.commands import bot_settings, debt_display, debt_management, user_settings
 from bot.utilities import send_messages, user_preferences
 
 # -------------------------------
@@ -137,9 +137,12 @@ class FakeAPI:
         self.calls['settle_debt'] = payload
         return {'settled_amount': payload['amount'], 'remaining_amount': '0'}
 
-    def set_unicode_preference(self, payload):
-        self.calls['set_unicode_preference'] = payload
-        return {'message': 'Preference updated'}
+    def set_unicode_preference(self, user_id, payload):
+        self.calls['set_unicode_preference'] = {
+            'user_id': user_id,
+            'use_unicode': payload['use_unicode']
+        }
+        return {"message": "Preference updated!"}
 
 # -------------------------------
 # Shared state between mocks
@@ -175,14 +178,14 @@ def mock_bot_dependencies(monkeypatch, shared):
 
     for module, model_names in {
         debt_management: ['OweRequest', 'SettleRequest'],
-        bot_settings: ['SetUnicodePreferenceRequest'],
+        user_settings: ['SetUnicodePreferenceRequest'],
     }.items():
         for name in model_names:
             monkeypatch.setattr(module, name, lambda **kwargs: type("DummyModel", (), {"model_dump": lambda self: kwargs})())
 
     monkeypatch.setattr(debt_management, 'api_client', shared.fake_api)
     monkeypatch.setattr(debt_display, 'api_client', shared.fake_api)
-    monkeypatch.setattr(bot_settings, 'api_client', shared.fake_api)
+    monkeypatch.setattr(user_settings, 'api_client', shared.fake_api)
 
 # -------------------------------
 # Intercept message functions
